test(ThemeAcceptanceTest): 테마 관련 인수테스트 추가
